The read class annotations are now correctly processed as an array of numerically indexed values in AnnotationDriver#loadMetadataForClass().
Since only one Document realted annotation should be present an exception is
thrown if more than one is found. (One could just use the first annotation,
but this may cause confusion when debugging.)

The type of the instanced annotation class is used to determine the type
of Document since there are no array keys to check against.